qmlui: preliminary implementation of the Show Manager
It is possible to:
- create a new Show
- edit the Show name
- drag existing Functions over the timeline including the creation of a new
Show/Track if needed

+++ engine/src/audio/audio.cpp
@@ -150,9 +150,9 @@ quint32 Audio::getStartTime() const
     return m_startTime;
 }
 
-qint64 Audio::totalDuration()
+quint32 Audio::totalDuration()
 {
-    return m_audioDuration;
+    return (quint32)m_audioDuration;
 }
 
 void Audio::setColor(QColor color)
